+++ Scripts/Maintenance/Get-AllProjectPublicationPackagesWithoutReferences.ps1
@@ -6,7 +6,7 @@ Set-MSConfigDebugMode -Value ($false); # $true oder $false
 #$WSMProjectGUID = ("8c681c01-ece4-4fc8-b1fb-45c4a7760501"); # Vodafone Intranet
 #$WSMProjectGUID = ("8A300F4B74AA4D4CB5C54C2165FC56D1"); # Shared Components - Simplicity
 #$WSMProjectGUID = ("6A89CD93325D42FD81A42FA836547016"); # vodafone Internet - Fachhandel Online
-#$WSMProjectGUID = ("7C503F8E943E4E6CACE63B2DEB82BBF4"); # Development - Defects and Testing
+$WSMProjectGUID = ("7C503F8E943E4E6CACE63B2DEB82BBF4"); # Development - Defects and Testing
 
 Register-MSSession -UseDefaults ($true);
 Select-MSSession -UseDefaults ($true);